+++ app/src/main/java/app/zornslemma/mypricelog/ui/screens/editprice/EditPriceScreen.kt
@@ -27,35 +27,54 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import app.zornslemma.mypricelog.R
+import app.zornslemma.mypricelog.common.formatPercent
 import app.zornslemma.mypricelog.data.DataSet
 import app.zornslemma.mypricelog.data.EditablePrice
 import app.zornslemma.mypricelog.data.Item
+import app.zornslemma.mypricelog.data.unitPrice
 import app.zornslemma.mypricelog.debug.myCheck
 import app.zornslemma.mypricelog.domain.MeasurementUnit
 import app.zornslemma.mypricelog.domain.QuantityType
+import app.zornslemma.mypricelog.domain.UnitPrice
 import app.zornslemma.mypricelog.domain.areDifferentUnitFamilies
+import app.zornslemma.mypricelog.domain.baseUnit
+import app.zornslemma.mypricelog.domain.calculateFriendlyUnitPriceDenominator
 import app.zornslemma.mypricelog.domain.getRelevantMeasurementUnits
+import app.zornslemma.mypricelog.ui.common.AsyncOperationStatus
 import app.zornslemma.mypricelog.ui.common.isNotBusy
 import app.zornslemma.mypricelog.ui.components.FilteredTextField
+import app.zornslemma.mypricelog.ui.components.LabeledItem
 import app.zornslemma.mypricelog.ui.components.MyExposedDropdownMenuBox
+import app.zornslemma.mypricelog.ui.components.SharedViewModel
+import app.zornslemma.mypricelog.ui.components.UnitPriceField
 import app.zornslemma.mypricelog.ui.components.ValidatedNumericTextField
 import app.zornslemma.mypricelog.ui.components.createOnCandidateValueChangeMaxLength
 import app.zornslemma.mypricelog.ui.components.generaledit.GeneralEditScreen
 import app.zornslemma.mypricelog.ui.components.keyboardCapitalization
 import app.zornslemma.mypricelog.ui.components.rememberSyncedTextFieldValue
+import app.zornslemma.mypricelog.ui.components.selectUnitPriceDenominator
 import app.zornslemma.mypricelog.ui.components.textOrNull
 import app.zornslemma.mypricelog.ui.components.topAppBarTitle
 import app.zornslemma.mypricelog.ui.maxNotesLength
 import app.zornslemma.mypricelog.ui.nonBreakingSpace
+import java.util.Locale
+import kotlin.math.abs
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun EditPriceScreen(viewModel: EditPriceViewModel, requestClose: (Long?) -> Unit) {
+fun EditPriceScreen(
+    sharedViewModel: SharedViewModel,
+    viewModel: EditPriceViewModel,
+    requestClose: (Long?) -> Unit,
+) {
     val originalPrice = viewModel.uiContent.originalContent
     val editablePrice by viewModel.uiContent.editableContent.collectAsStateWithLifecycle()
     val dataSet = viewModel.uiContent.staticContent.dataSet
     val item = viewModel.uiContent.staticContent.item
     val source = viewModel.uiContent.staticContent.source
+    val autoUnitPriceDenominator = viewModel.uiContent.staticContent.autoUnitPriceDenominator
+    val nonLinearEdit = viewModel.uiContent.staticContent.nonLinearEdit
+    val locale = viewModel.uiContent.staticContent.frozenLocale
 
     val saveStatus by
         viewModel.generalEditScreenStateHolder.asyncOperationStatus.collectAsStateWithLifecycle()
@@ -105,6 +124,19 @@ fun EditPriceScreen(viewModel: EditPriceViewModel, requestClose: (Long?) -> Unit
         Spacer(modifier = Modifier.height(16.dp))
 
         EditPriceScreenPackSize(viewModel, editablePrice, dataSet, item, ::onPackSizeOrPriceChange)
+
+        Spacer(modifier = Modifier.height(16.dp))
+
+        EditPriceScreenUnitPriceRow(
+            sharedViewModel,
+            originalPrice,
+            editablePrice,
+            dataSet,
+            autoUnitPriceDenominator,
+            nonLinearEdit,
+            locale,
+            saveStatus,
+        )
 
         // We don't show the switch if this is the first price for an item and source; the price is
         // confirmed, otherwise why are we entering it? Note that this is not the same as id being
@@ -330,6 +362,135 @@ private fun EditPriceScreenPackSize(
                     areDifferentUnitFamilies(previousItem, item)
                 },
             )
+        }
+    }
+}
+
+@Composable
+fun EditPriceScreenUnitPriceRow(
+    sharedViewModel: SharedViewModel,
+    originalPrice: EditablePrice,
+    editablePrice: EditablePrice,
+    dataSet: DataSet,
+    autoUnitPriceDenominator: MeasurementUnit,
+    nonLinearEdit: Boolean,
+    locale: Locale,
+    saveStatus: AsyncOperationStatus,
+) {
+    // ENHANCE: It's possible we should use the three-field layout even when we're adding a new
+    // price and there is no old price or percentage change to show - just leaving the unused
+    // fields empty or with a "-" or something. It seems a bit silly to complicate the UI with these
+    // redundant fields, but it may look better in practice (the single unit price field looks a
+    // bit of an orphan) and there may be a consistency argument too.
+
+    // ENHANCE: Should we have a slight delay in updating the unit price when the user is typing
+    // into the contributing text fields? It might be less distracting, but it might also make the
+    // app feel laggy.
+
+    Row(modifier = Modifier.fillMaxWidth()) {
+        val userUnitPriceDenominator by
+            sharedViewModel.userUnitPriceDenominatorFlow.collectAsStateWithLifecycle()
+
+        // If we're adding a new price we obviously don't have an original unit price. We also
+        // don't really have one if this is a non-linear edit. We could in theory take the
+        // current price if there is one, but I'm not sure it's particularly useful.
+        // originalPrice is the starting copy of the historical price, which feels confusing to
+        // show here.
+        val haveOriginalUnitPrice = originalPrice.id != 0L && !nonLinearEdit
+
+        val originalUnitPrice =
+            if (!haveOriginalUnitPrice) null
+            else
+                originalPrice
+                    .unitPrice(locale)
+                    ?.withDenominator(
+                        selectUnitPriceDenominator(
+                            autoUnitPriceDenominator,
+                            userUnitPriceDenominator,
+                        )
+                    )
+        val newUnitPrice = editablePrice.unitPrice(locale)
+        val currentUnitPriceDenominator =
+            remember(
+                userUnitPriceDenominator,
+                originalUnitPrice,
+                newUnitPrice,
+                dataSet,
+                editablePrice.measurementUnit,
+            ) {
+                if (userUnitPriceDenominator != null) userUnitPriceDenominator!!
+                else {
+                    originalUnitPrice?.denominator
+                        ?: if (newUnitPrice != null) {
+                            calculateFriendlyUnitPriceDenominator(
+                                dataSet,
+                                editablePrice.measurementUnit,
+                                listOf(newUnitPrice),
+                            )
+                        } else {
+                            null
+                        }
+                }
+            }
+
+        @Composable
+        fun LocalUnitPriceField(unitPrice: UnitPrice?, label: String) {
+            UnitPriceField(
+                unitPrice = unitPrice,
+                dataSet = dataSet,
+                enabled = saveStatus.isNotBusy(),
+                label = label,
+                modifier = Modifier.weight(3f),
+                onItemSelected = { sharedViewModel.updateUserUnitPriceDenominator(it) },
+            )
+        }
+
+        if (originalUnitPrice != null) {
+            LocalUnitPriceField(
+                unitPrice = originalUnitPrice.withDenominator(currentUnitPriceDenominator!!),
+                label = stringResource(R.string.label_existing_unit_price),
+            )
+        }
+
+        LocalUnitPriceField(
+            unitPrice = newUnitPrice?.withDenominator(currentUnitPriceDenominator!!),
+            label =
+                stringResource(
+                    if (originalUnitPrice == null) R.string.label_unit_price
+                    else R.string.label_new_unit_price
+                ),
+        )
+
+        if (originalUnitPrice != null) {
+            LabeledItem(
+                label = stringResource(R.string.label_change),
+                modifier = Modifier.weight(2f),
+            ) {
+                val baseUnit = originalPrice.measurementUnit.quantityType.baseUnit()
+                if (newUnitPrice == null) {
+                    Text("")
+                } else {
+                    val ratio =
+                        newUnitPrice.withDenominator(baseUnit).numerator /
+                            originalUnitPrice!!.withDenominator(baseUnit).numerator
+                    val relativeChange = (ratio - 1)
+                    // ENHANCE: It might be nice to colour code the percentage change, e.g. red for
+                    // positive (bad) and green for negative (good). However, elsewhere we would
+                    // have liked to use red/green in this way but it didn't feel like a good fit
+                    // with Material You. We could use the "error" colour for bad and the primary
+                    // colour for good, but maybe this is confusing.
+                    Text(
+                        formatPercent(
+                            relativeChange,
+                            1,
+                            1,
+                            true,
+                            locale,
+                            useLeadingPlus = abs(relativeChange) >= 0.0005,
+                        )
+                    )
+                }
+            }
         }
     }
 }